backend/pyproject.toml
[project]
name = "careops-backend"
version = "0.1.0"
description = "CareOps Unified Operations Platform — Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.18",
    "greenlet>=3.1.0",
    "httpx>=0.27.0",
    "slowapi>=0.1.9",
    "aiosmtplib>=2.0.0",
    "cryptography>=46.0.4",
    "twilio>=9.10.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

